new update

+++ src/controllers/consultant.controller.js
@@ -1,11 +1,11 @@
-const { success, failed } = require('../helpers/response');
+const { success, failed } = require("../helpers/response");
 const {
   getConsultantList,
   getConsultantDetail,
   updateConsultantProfile,
   updateConsultantPhoto
-} = require('../models/consultant.model');
-const fs = require('fs');
+} = require("../models/consultant.model");
+const fs = require("fs");
 
 module.exports = {
   getConsultantList: async (req, res) => {
@@ -20,7 +20,7 @@ module.exports = {
         });
       }
 
-      return failed(res, 400, 'No data found');
+      return failed(res, 400, "No data found");
     } catch (error) {
       return failed(res, 500, {
         message: error.message
@@ -39,7 +39,7 @@ module.exports = {
         });
       }
 
-      return failed(res, 400, 'No data found');
+      return failed(res, 400, "No data found");
     } catch (error) {
       return failed(res, 500, {
         message: error.message
@@ -66,11 +66,11 @@ module.exports = {
 
       if (result.affectedRows > 0) {
         return success(res, 200, {
-          message: 'Successfully updated profile'
+          message: "Successfully updated profile"
         });
       }
 
-      return failed(res, 400, 'Failed to update profile');
+      return failed(res, 400, "Failed to update profile");
     } catch (error) {
       return failed(res, 500, {
         message: error.message
@@ -82,9 +82,22 @@ module.exports = {
     try {
       const { id } = req.params;
       const profile = await getConsultantDetail({ consultantId: id });
+
+      if (profile.length === 0) {
+        return failed(res, 400, "Consultant not found");
+      }
+
+      if (profile[0].userId !== req.userData.userId) {
+        return failed(
+          res,
+          400,
+          "You are not authorized to update this profile"
+        );
+      }
+
       const { photo } = profile[0];
 
-      if (photo !== 'default.png') {
+      if (photo !== "default.png") {
         fs.unlinkSync(`./public/${photo}`);
       }
 
@@ -96,11 +109,11 @@ module.exports = {
 
       if (result.affectedRows > 0) {
         return success(res, 200, {
-          message: 'Profile photo updated'
+          message: "Profile photo updated"
         });
       }
 
-      return failed(res, 400, 'failed to update profile photo');
+      return failed(res, 400, "failed to update profile photo");
     } catch (error) {
       return failed(res, 500, {
         message: error.message
